+++ Common/src/main/java/de/markusbordihn/easynpc/entity/easynpc/npc/Villager.java
@@ -23,6 +23,7 @@ import de.markusbordihn.easynpc.data.skin.SkinModel;
 import de.markusbordihn.easynpc.data.sound.SoundDataSet;
 import de.markusbordihn.easynpc.data.sound.SoundType;
 import de.markusbordihn.easynpc.entity.EasyNPCBaseModelEntity;
+import de.markusbordihn.easynpc.entity.easynpc.npc.Chicken.Variant;
 import de.markusbordihn.easynpc.network.components.TextComponent;
 import de.markusbordihn.easynpc.utils.TextUtils;
 import net.minecraft.network.chat.Component;
@@ -113,7 +114,11 @@ public class Villager extends EasyNPCBaseModelEntity<Villager> {
 
   @Override
   public Enum<?> getVariant(String name) {
-    return Variant.valueOf(name);
+    try {
+      return Variant.valueOf(name);
+    } catch (IllegalArgumentException e) {
+      return getDefaultVariant();
+    }
   }
 
   @Override