+++ Fabric/src/main/java/de/markusbordihn/easynpc/entity/ModEntityType.java
@@ -40,6 +40,8 @@ import de.markusbordihn.easynpc.entity.easynpc.npc.Villager;
 import de.markusbordihn.easynpc.entity.easynpc.npc.Wolf;
 import de.markusbordihn.easynpc.entity.easynpc.npc.Zombie;
 import de.markusbordihn.easynpc.entity.easynpc.npc.ZombieVillager;
+import de.markusbordihn.easynpc.entity.easynpc.raw.PiglinRaw;
+import de.markusbordihn.easynpc.entity.easynpc.raw.SkeletonRaw;
 import de.markusbordihn.easynpc.entity.easynpc.raw.ZombieRaw;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
 import net.minecraft.core.Registry;
@@ -155,11 +157,21 @@ public class ModEntityType {
           Registry.ENTITY_TYPE,
           Constants.MOD_ID + ":" + ZombieVillager.ID,
           ModEntityTypes.ZOMBIE_VILLAGER);
+
   // Raw entities (for modding only)
+  public static final EntityType<PiglinRaw> PIGLIN_RAW =
+      Registry.register(
+          Registry.ENTITY_TYPE, Constants.MOD_ID + ":" + PiglinRaw.ID, ModEntityTypes.PIGLIN_RAW);
+  public static final EntityType<SkeletonRaw> SKELETON_RAW =
+      Registry.register(
+          Registry.ENTITY_TYPE,
+          Constants.MOD_ID + ":" + SkeletonRaw.ID,
+          ModEntityTypes.SKELETON_RAW);
   public static final EntityType<ZombieRaw> ZOMBIE_RAW =
       Registry.register(
           Registry.ENTITY_TYPE, Constants.MOD_ID + ":" + ZombieRaw.ID, ModEntityTypes.ZOMBIE_RAW);
   private static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
+
   // Optional: Epic Fight entities
   public static EntityType<EpicFightZombie> EPIC_FIGHT_ZOMBIE;
 
@@ -205,6 +217,10 @@ public class ModEntityType {
     FabricDefaultAttributeRegistry.register(ZOMBIE_VILLAGER, ZombieVillager.createAttributes());
 
     // Raw entities (for modding only)
+    FabricDefaultAttributeRegistry.register(
+        PIGLIN_RAW, net.minecraft.world.entity.monster.piglin.Piglin.createAttributes());
+    FabricDefaultAttributeRegistry.register(
+        SKELETON_RAW, net.minecraft.world.entity.monster.Skeleton.createAttributes());
     FabricDefaultAttributeRegistry.register(
         ZOMBIE_RAW, net.minecraft.world.entity.monster.Zombie.createAttributes());